Updates git-repositories lookup in drafts
